fix(mtf): Fix signed/unsigned integer comparison warning -- src/mt/mtf/mtfDecodeOffset.cpp
Cast sValue to UInt before comparison with (sFence - 1) to resolve compiler warning.
This is safe because sValue is verified to be positive (> 0) in the preceding check,
ensuring no data loss during the cast from signed to unsigned integer.

Note:
[ fix warnings ]

src/mt/mtf/mtfDecodeOffset.cpp:311:53: warning: comparison of integers of different signs: 'SInt' (aka 'int') and 'UInt' (aka 'unsigned int') [-Wsign-compare]
  311 |         IDE_TEST_RAISE( ( sValue <= 0 ) || ( sValue > ( sFence - 1 ) ),
      |                                              ~~~~~~ ^   ~~~~~~~~~~
src/id/include/ideErrorMgr.h:600:57: note: expanded from macro 'IDE_TEST_RAISE'
  600 | #define IDE_TEST_RAISE( a, b ) { if ( IDL_LIKELY_FALSE( a ) ) {         \
      |                                                         ^
src/id/include/idl.h:835:62: note: expanded from macro 'IDL_LIKELY_FALSE'
  835 | #define IDL_LIKELY_FALSE(x)      __builtin_expect((long int)(x), 0)
      |                                                              ^

+++ src/mt/mtf/mtfDecodeOffset.cpp
@@ -308,7 +308,7 @@ IDE_RC mtfDecodeOffsetCalculate( mtcNode*     aNode,
         }
 
         // 범위를 넘으면 에러
-        IDE_TEST_RAISE( ( sValue <= 0 ) || ( sValue > ( sFence - 1 ) ),
+        IDE_TEST_RAISE( ( sValue <= 0 ) || ( (UInt)sValue > ( sFence - 1 ) ),
                          ERR_ARGUMENT_NOT_APPLICABLE );
 
         // Calculate the Nth( offset_value ) node
